+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -38,6 +38,7 @@
 #include "Rule19p2Check.h"
 #include "Rule8p10Check.h"
 #include "Rule8p14Check.h"
+#include "Rule8p2Check.h"
 #include "Rule9p3Check.h"
 
 namespace clang::tidy {
@@ -106,6 +107,8 @@ public:
         "misrac-rule-8p10");
     CheckFactories.registerCheck<Rule8p14Check>(
         "misrac-rule-8p14");
+    CheckFactories.registerCheck<Rule8p2Check>(
+        "misrac-rule-8p2");
     CheckFactories.registerCheck<Rule9p3Check>(
         "misrac-rule-9p3");
   }