+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -377,6 +377,12 @@ New checks
   Detects issues in accordance with MISRA C rule 19.2:
   "`The union keyword should not be used`".
 
+- New :doc:`misrac-rule-8p2
+  <clang-tidy/checks/misrac/rule-8p2>` check.
+
+  Detects issues in accordance with MISRA C rule 8.2:
+  "`Function types shall be in prototype form with named parameters`".
+
 - New :doc:`misrac-rule-8p10
   <clang-tidy/checks/misrac/rule-8p10>` check.